Initial commit
Changed the names of the css and html files. Added an form. The text area is in the form and the id of the text area is sampleBox. the name  of the text area is sampleText. Added autofocus for the text area. Added the placeholder for the text area. Added the field set and the field set has an legend of Fonts.